+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- changed: Rename Thorchain DEX Aggregator to SwapKit
+
 ## 2.13.0 (2024-10-31)
 
 - added: Add TON unique IDs to swap partners